feat: Правки по результатам ревью
