+++ src/handlers/EpicHandler.java
@@ -13,76 +13,46 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 public class EpicHandler extends BaseHttpHandler {
-    private final TaskManager taskManager;
-
     public EpicHandler(TaskManager taskManager) {
-        this.taskManager = taskManager;
+        super(taskManager);
     }
 
     @Override
-    public void handle(HttpExchange exchange) throws IOException {
+    protected void handleGetRequest(HttpExchange exchange) throws IOException {
         String[] urlElements = getURLElements(exchange);
         String lasUrlElement = urlElements[urlElements.length - 1];
         String penultimateElement = urlElements[urlElements.length - 2];
-        switch (getMethod(exchange)) {
-            case "GET": {
-                if (hasIdInUrl(lasUrlElement)) {
-                    getEpic(exchange, Integer.parseInt(lasUrlElement));
-                } else if (lasUrlElement.equals("subtasks") && hasIdInUrl(penultimateElement)) {
-                    getEpicSubtasks(exchange, Integer.parseInt(penultimateElement));
-                } else {
-                    if (lasUrlElement.equals("epics")) {
-                        getAllEpics(exchange);
-                    } else {
-                        sendNotFound(exchange, null);
-                    }
-                }
-                break;
+        if (hasIdInUrl(lasUrlElement)) {
+            int id = Integer.parseInt(lasUrlElement);
+            try {
+                Epic epic = taskManager.getEpic(id);
+                sendText(exchange, gson.toJson(epic));
+            } catch (TaskNotFound e) {
+                sendNotFound(exchange, String.format("Epic с id %d не найден", id));
             }
-            case "POST": {
-                createOrUpdateEpic(exchange);
-                break;
+        } else if (lasUrlElement.equals("subtasks") && hasIdInUrl(penultimateElement)) {
+            int id = Integer.parseInt(penultimateElement);
+            try {
+                Epic epic = taskManager.getEpic(id);
+                ArrayList<Subtask> subtasks = epic.getSubtasks();
+                String json = gson.toJson(subtasks);
+                sendText(exchange, json);
+            } catch (TaskNotFound e) {
+                sendNotFound(exchange, String.format("Epic с id %d не найден", id));
             }
-            case "DELETE": {
-                if (hasIdInUrl(getLastURLElement(exchange))) {
-                    deleteEpic(exchange, Integer.parseInt(lasUrlElement));
-                } else {
-                    sendNotFound(exchange, null);
-                }
-                break;
+        } else {
+            if (lasUrlElement.equals("epics")) {
+                ArrayList<Epic> epics = taskManager.getEpics();
+                String json = gson.toJson(epics);
+                sendText(exchange, json);
+            } else {
+                sendNotFound(exchange, null);
             }
-            default:
-                sendForbiddenMethod(exchange);
         }
     }
 
-    public void getEpic(HttpExchange exchange, int id) throws IOException {
-        try {
-            Epic epic = taskManager.getEpic(id);
-            sendText(exchange, gson.toJson(epic));
-        } catch (TaskNotFound e) {
-            sendNotFound(exchange, String.format("Epic с id %d не найден", id));
-        }
-    }
-
-    public void getEpicSubtasks(HttpExchange exchange, int id) throws IOException {
-        try {
-            Epic epic = taskManager.getEpic(id);
-            ArrayList<Subtask> subtasks = epic.getSubtasks();
-            String json = gson.toJson(subtasks);
-            sendText(exchange, json);
-        } catch (TaskNotFound e) {
-            sendNotFound(exchange, String.format("Epic с id %d не найден", id));
-        }
-    }
-
-    public void getAllEpics(HttpExchange exchange) throws IOException {
-        ArrayList<Epic> epics = taskManager.getEpics();
-        String json = gson.toJson(epics);
-        sendText(exchange, json);
-    }
-
-    public void createOrUpdateEpic(HttpExchange exchange) throws IOException {
+    @Override
+    protected void handlePostRequest(HttpExchange exchange) throws IOException {
         JsonObject taskJson = getJsonObject(exchange);
         JsonElement id = taskJson.get("taskId");
         Epic epic = gson.fromJson(taskJson, Epic.class);
@@ -102,12 +72,18 @@ public class EpicHandler extends BaseHttpHandler {
         }
     }
 
-    public void deleteEpic(HttpExchange exchange, int epicId) throws IOException {
-        try {
-            taskManager.deleteEpic(epicId);
-            sendText(exchange, String.format("Epic с id %d удален", epicId));
-        } catch (TaskNotFound e) {
-            sendNotFound(exchange, String.format("Epic с id %d не найден", epicId));
+    @Override
+    protected void handleDeleteRequest(HttpExchange exchange) throws IOException {
+        if (hasIdInUrl(getLastURLElement(exchange))) {
+            int epicId = Integer.parseInt(getLastURLElement(exchange));
+            try {
+                taskManager.deleteEpic(epicId);
+                sendText(exchange, String.format("Epic с id %d удален", epicId));
+            } catch (TaskNotFound e) {
+                sendNotFound(exchange, String.format("Epic с id %d не найден", epicId));
+            }
+        } else {
+            sendNotFound(exchange, null);
         }
     }
 }
