+++ src/handlers/HistoryHandler.java
@@ -6,22 +6,17 @@ import controller.TaskManager;
 import java.io.IOException;
 
 public class HistoryHandler extends BaseHttpHandler {
-    private final TaskManager taskManager;
 
     public HistoryHandler(TaskManager taskManager) {
-        this.taskManager = taskManager;
+        super(taskManager);
     }
 
     @Override
-    public void handle(HttpExchange exchange) throws IOException {
-        if (getMethod(exchange).equals("GET")) {
-            if (getLastURLElement(exchange).equals("history")) {
-                sendText(exchange, gson.toJson(taskManager.getHistory()));
-            } else {
-                sendNotFound(exchange, null);
-            }
+    protected void handleGetRequest(HttpExchange exchange) throws IOException {
+        if (getLastURLElement(exchange).equals("history")) {
+            sendText(exchange, gson.toJson(taskManager.getHistory()));
         } else {
-            sendForbiddenMethod(exchange);
+            sendNotFound(exchange, null);
         }
     }
 }
