+++ src/handlers/PrioritizedHandler.java
@@ -6,22 +6,17 @@ import controller.TaskManager;
 import java.io.IOException;
 
 public class PrioritizedHandler extends BaseHttpHandler {
-    TaskManager taskManager;
 
     public PrioritizedHandler(TaskManager taskManager) {
-        this.taskManager = taskManager;
+        super(taskManager);
     }
 
     @Override
-    public void handle(HttpExchange exchange) throws IOException {
-        if (getMethod(exchange).equals("GET")) {
-            if (getLastURLElement(exchange).equals("prioritized")) {
-                sendText(exchange, gson.toJson(taskManager.getPrioritizedTasks()));
-            } else {
-                sendNotFound(exchange, null);
-            }
+    protected void handleGetRequest(HttpExchange exchange) throws IOException {
+        if (getLastURLElement(exchange).equals("prioritized")) {
+            sendText(exchange, gson.toJson(taskManager.getPrioritizedTasks()));
         } else {
-            sendForbiddenMethod(exchange);
+            sendNotFound(exchange, null);
         }
     }
 }
