+++ src/handlers/TaskHandler.java
@@ -13,61 +13,34 @@ import java.io.IOException;
 import java.util.ArrayList;
 
 public class TaskHandler extends BaseHttpHandler {
-    private final TaskManager taskManager;
-
     public TaskHandler(TaskManager taskManager) {
-        this.taskManager = taskManager;
+        super(taskManager);
     }
 
     @Override
-    public void handle(HttpExchange exchange) throws IOException {
+    protected void handleGetRequest(HttpExchange exchange) throws IOException {
         String lasUrlElement = getLastURLElement(exchange);
-        switch (getMethod(exchange)) {
-            case "GET": {
-                if (hasIdInUrl(lasUrlElement)) {
-                    getTask(exchange, Integer.parseInt(lasUrlElement));
-                } else {
-                    if (lasUrlElement.equals("tasks")) {
-                        getAllTasks(exchange);
-                    } else {
-                        sendNotFound(exchange, null);
-                    }
-                }
-                break;
+        if (hasIdInUrl(lasUrlElement)) {
+            int taskId = Integer.parseInt(lasUrlElement);
+            try {
+                Task task = taskManager.getTask(taskId);
+                sendText(exchange, gson.toJson(task));
+            } catch (TaskNotFound e) {
+                sendNotFound(exchange, String.format("Задача с id %d не найдена", taskId));
             }
-            case "POST": {
-                createOrUpdateTask(exchange);
-                break;
+        } else {
+            if (lasUrlElement.equals("tasks")) {
+                ArrayList<Task> tasks = taskManager.getTasks();
+                String json = gson.toJson(tasks);
+                sendText(exchange, json);
+            } else {
+                sendNotFound(exchange, null);
             }
-            case "DELETE": {
-                if (hasIdInUrl(lasUrlElement)) {
-                    deleteTask(exchange, Integer.parseInt(lasUrlElement));
-                } else {
-                    sendNotFound(exchange, null);
-                }
-                break;
-            }
-            default:
-                sendForbiddenMethod(exchange);
         }
     }
 
-    public void getTask(HttpExchange exchange, int id) throws IOException {
-        try {
-            Task task = taskManager.getTask(id);
-            sendText(exchange, gson.toJson(task));
-        } catch (TaskNotFound e) {
-            sendNotFound(exchange, String.format("Задача с id %d не найдена", id));
-        }
-    }
-
-    public void getAllTasks(HttpExchange exchange) throws IOException {
-        ArrayList<Task> tasks = taskManager.getTasks();
-        String json = gson.toJson(tasks);
-        sendText(exchange, json);
-    }
-
-    public void createOrUpdateTask(HttpExchange exchange) throws IOException {
+    @Override
+    protected void handlePostRequest(HttpExchange exchange) throws IOException {
         JsonObject taskJson = getJsonObject(exchange);
         JsonElement id = taskJson.get("taskId");
         Task task = gson.fromJson(taskJson, Task.class);
@@ -87,12 +60,19 @@ public class TaskHandler extends BaseHttpHandler {
         }
     }
 
-    public void deleteTask(HttpExchange exchange, int taskId) throws IOException {
-        try {
-            taskManager.deleteTask(taskId);
-            sendText(exchange, String.format("Задача с id %d удалена", taskId));
-        } catch (TaskNotFound e) {
-            sendNotFound(exchange, String.format("Задача с id %d не найдена", taskId));
+    @Override
+    protected void handleDeleteRequest(HttpExchange exchange) throws IOException {
+        String lasUrlElement = getLastURLElement(exchange);
+        if (hasIdInUrl(lasUrlElement)) {
+            int taskId = Integer.parseInt(lasUrlElement);
+            try {
+                taskManager.deleteTask(taskId);
+                sendText(exchange, String.format("Задача с id %d удалена", taskId));
+            } catch (TaskNotFound e) {
+                sendNotFound(exchange, String.format("Задача с id %d не найдена", taskId));
+            }
+        } else {
+            sendNotFound(exchange, null);
         }
     }
 }
